refactor: Separate the tree a transform binds against from the edited tree
A SemanticModel only answers for nodes of the tree it was created from, so the
transform path has to name which tree it binds against before any declaration
can be removed from it. Each source unit now carries a binding tree beside the
edited one, and every consumer that pairs a root with the semantic model reads
the binding root: the emit planner, the property getter emitter, and the const
and member-kind drift checks. The edited root keeps the diagnostic positions and
the line directives, and the unannotated root stays the baseline's.

The binding tree is the loaded tree until a caller removes something, so this
changes no behavior on its own.

Fingerprint computation also moves out of the planner into its own type, because
the declaration verifier has to recompute a fingerprint through the same path.

Verification: every TransformWorker test class, 244 passed / 0 failed, single
flight.
